1，重构loggingHandler模块 2，重构timeout模块 3，调整相应的core的example 4，UserEvent使用boost::any替代

+++ cetty-core/src/cetty/logging/LogPatternFormatter.cpp
@@ -48,7 +48,7 @@ void LogPatternFormatter::format(LogMessage& msg, const std::string& pattern) co
         if (*it == '%') {
             if (++it != end) {
                 switch (*it) {
-                case 'l': msg.getStream() << msg.getLevel(); break;
+                case 'l': msg.getStream() << msg.getLevel().toInt(); break;
 
                 case 'L': msg.getStream() << msg.getLevel().toString(); break;
